factor WhereTrait into WhereInterface
