+++ src/Common/Delete.php
@@ -22,8 +22,6 @@ use Aura\Sql_Query\Traits;
  */
 class Delete extends AbstractQuery implements DeleteInterface
 {
-    use Traits\WhereTrait;
-    
     /**
      *
      * The table to delete from.
@@ -74,5 +72,51 @@ class Delete extends AbstractQuery implements DeleteInterface
     protected function buildFrom()
     {
         $this->stm .= " FROM {$this->from}";
+    }
+
+    /**
+     *
+     * Adds a WHERE condition to the query by AND. If the condition has
+     * ?-placeholders, additional arguments to the method will be bound to
+     * those placeholders sequentially.
+     *
+     * @param string $cond The WHERE condition.
+     * @param mixed ...$bind arguments to bind to placeholders
+     *
+     * @return $this
+     *
+     */
+    public function where($cond)
+    {
+        $bind = func_get_args();
+        array_shift($bind);
+
+        $this->addWhere($cond, 'AND', $bind);
+
+        return $this;
+    }
+
+    /**
+     *
+     * Adds a WHERE condition to the query by OR. If the condition has
+     * ?-placeholders, additional arguments to the method will be bound to
+     * those placeholders sequentially.
+     *
+     * @param string $cond The WHERE condition.
+     * @param mixed ...$bind arguments to bind to placeholders
+     *
+     * @return $this
+     *
+     * @see where()
+     *
+     */
+    public function orWhere($cond)
+    {
+        $bind = func_get_args();
+        array_shift($bind);
+
+        $this->addWhere($cond, 'OR', $bind);
+
+        return $this;
     }
 }
