+++ src/Common/Select.php
@@ -25,7 +25,6 @@ class Select extends AbstractQuery implements SelectInterface
 {
     use Traits\LimitOffsetTrait;
     use Traits\OrderByTrait;
-    use Traits\WhereTrait;
 
     /**
      * 
@@ -603,5 +602,51 @@ class Select extends AbstractQuery implements SelectInterface
         if ($this->for_update) {
             $this->stm .= PHP_EOL . 'FOR UPDATE';
         }
+    }
+
+    /**
+     *
+     * Adds a WHERE condition to the query by AND. If the condition has
+     * ?-placeholders, additional arguments to the method will be bound to
+     * those placeholders sequentially.
+     *
+     * @param string $cond The WHERE condition.
+     * @param mixed ...$bind arguments to bind to placeholders
+     *
+     * @return $this
+     *
+     */
+    public function where($cond)
+    {
+        $bind = func_get_args();
+        array_shift($bind);
+
+        $this->addWhere($cond, 'AND', $bind);
+
+        return $this;
+    }
+
+    /**
+     *
+     * Adds a WHERE condition to the query by OR. If the condition has
+     * ?-placeholders, additional arguments to the method will be bound to
+     * those placeholders sequentially.
+     *
+     * @param string $cond The WHERE condition.
+     * @param mixed ...$bind arguments to bind to placeholders
+     *
+     * @return $this
+     *
+     * @see where()
+     *
+     */
+    public function orWhere($cond)
+    {
+        $bind = func_get_args();
+        array_shift($bind);
+
+        $this->addWhere($cond, 'OR', $bind);
+
+        return $this;
     }
 }
