+++ src/Common/WhereInterface.php
@@ -1,0 +1,45 @@
+<?php
+
+namespace Aura\Sql_Query\Common;
+
+interface WhereInterface
+{
+    /**
+     *
+     * Adds a WHERE condition to the query by AND. If the condition has
+     * ?-placeholders, additional arguments to the method will be bound to
+     * those placeholders sequentially.
+     *
+     * @param string $cond The WHERE condition.
+     * @param mixed ...$params arguments to be bound to placeholders
+     *
+     * @return $this
+     *
+     */
+    public function where($cond);
+
+    /**
+     *
+     * Adds a WHERE condition to the query by OR. If the condition has
+     * ?-placeholders, additional arguments to the method will be bound to
+     * those placeholders sequentially.
+     *
+     * @param string $cond The WHERE condition.
+     * @param mixed ...$params arguments to be bound to placeholders
+     *
+     * @return $this
+     *
+     * @see where()
+     *
+     */
+    public function orWhere($cond);
+
+    /**
+     *
+     * Gets the values to bind to placeholders.
+     *
+     * @return array
+     *
+     */
+    public function getBindValues();
+}
